requirements.txt
python-dotenv
streamlit
pytube
assemblyai
sentence-transformers
ctransformers
faiss-cpu
streamlit_player
replicate
